# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyExp API",
        "description": "API documentation for MyExp application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tickets/generate/{reservation_id}": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Generate ticket",
                "operationId": "805d5ef17bf9f2b2f1a7243fdd0a5666",
                "parameters": [
                    {
                        "name": "reservation_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                }
            }
        },
        "/api/ticket/{ticket_id}/pdf": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Generate and download ticket PDF",
                "description": "Generate a PDF file for a ticket including QR code and reservation details",
                "operationId": "bcea938e87db24bf401f5d6fde521732",
                "parameters": [
                    {
                        "name": "ticket_id",
                        "in": "path",
                        "description": "ID of the ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file downloaded successfully"
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}